requirements-web.txt
# 現有核心依賴
numpy>=1.24.0
pandas>=2.0.0
plotly>=5.0.0
backtesting>=0.6.0
python-dotenv>=1.0.0
requests>=2.28.0
python-binance>=1.0.0
alpaca-py>=0.39.0
ccxt>=4.0.0
pandas-ta>=0.3.14b0
pydantic>=2.0.0

# Web 服務器相關依賴
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.20
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
websockets>=15.0.0

# 前端模板和靜態文件
jinja2>=3.1.0

# 圖表和數據可視化
mplfinance>=0.12.10b0
bokeh>=3.7.0

# 數據庫支持 (可選)
sqlalchemy>=2.0.0
alembic>=1.13.0

# 日誌和監控
loguru>=0.7.0

# 開發工具
pytest>=7.0.0
pytest-asyncio>=0.23.0
black>=23.0.0
